--- openpathsampling/dynamics/openmm/features/__init__.py
@@ -1,3 +1,0 @@
-from openpathsampling.dynamics.features import *
-
-from openpathsampling.dynamics.features.shared import StaticContainer, KineticContainer
