+++ openpathsampling/engines/features/base.py
