+++ openpathsampling/engines/features/coordinates.py
