+++ openpathsampling/engines/features/engine.py
