+++ openpathsampling/engines/features/kinetics.py
