+++ openpathsampling/engines/features/numpydoctools.py
