+++ openpathsampling/engines/features/shared.py
