+++ openpathsampling/engines/features/statics.py
