+++ openpathsampling/engines/openmm/features/__init__.py
@@ -1,0 +1,3 @@
+from openpathsampling.engines.features import *
+
+from openpathsampling.engines.features.shared import StaticContainer, KineticContainer
