+++ openpathsampling/engines/snapshot.py
