+++ openpathsampling/engines/topology.py
