+++ openpathsampling/engines/toy/features/__init__.py
@@ -1,4 +1,4 @@
-from openpathsampling.dynamics.features import (
+from openpathsampling.engines.features import (
     base,
     coordinates,
     velocities,
