+++ openpathsampling/engines/trajectory.py
